+++ src/codec/UploadDarFileRequestCodec.ts
@@ -1,0 +1,20 @@
+// Copyright (c) 2019 The DAML Authors. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import {Codec} from "./Codec";
+import {UploadDarFileRequest} from "../model/UploadDarFileRequest";
+import {UploadDarFileRequest as PbUploadDarFileRequest} from "../generated/com/digitalasset/ledger/api/v1/admin/package_management_service_pb"
+
+export const UploadDarFileRequestCodec: Codec<PbUploadDarFileRequest, UploadDarFileRequest> = {
+    deserialize(message: PbUploadDarFileRequest): UploadDarFileRequest {
+        const object: UploadDarFileRequest = {
+            darFile: message.getDarFile_asB64()
+        };
+        return object;
+    },
+    serialize(object: UploadDarFileRequest): PbUploadDarFileRequest{
+        const message = new PbUploadDarFileRequest();
+        message.setDarFile(object.darFile);
+        return message;
+    }
+}